move methods from venta to another classes
