+++ Ventas/src/main/java/Product.java
@@ -1,0 +1,12 @@
+public class Product {
+
+    public static void getTotalPrice(int totalPriceHamburgesasBought, int totalPricePerrosBought) {
+        int totalPrice = totalPriceHamburgesasBought + totalPricePerrosBought;
+        System.out.println("Total: " + totalPrice);
+    }
+
+    public static int sellProduct(int price, int cuantity){
+        return price*cuantity;
+    }
+
+}
