+++ Ventas/src/main/java/main.java
@@ -12,8 +12,8 @@ public class main {
         System.out.println("Venta de perros y hamburguesas");
 
         while (takeOrder) {
-            productSelecction = productToSell(takeOrders);
-            cuantityProductSelection = cuantityToSell(takeOrders);
+            productSelecction = TakeOrders.productToSell(takeOrders);
+            cuantityProductSelection = TakeOrders.cuantityToSell(takeOrders);
             if (productSelecction == 1) {
                 perroCuantityToBuy +=  cuantityProductSelection;
             } else if (productSelecction== 2) {
@@ -21,7 +21,7 @@ public class main {
             } else {
                 break;
             }
-            endSale = takeMoreOrders(takeOrders);
+            endSale = TakeOrders.takeMoreOrders(takeOrders);
             if (endSale == 1) {
                 takeOrder = true;
             } else if (endSale == 2) {
@@ -31,33 +31,9 @@ public class main {
             }
         }
 
-        totalPriceHamburgesasBought = sellProduct(hamburguesaPrice,hamburguesaCuantityToBuy);
-        totalPricePerrosBought = sellProduct(perroPrice,perroCuantityToBuy);
-        getTotalPrice(totalPriceHamburgesasBought,totalPricePerrosBought);
-    }
-
-    private static void getTotalPrice(int totalPriceHamburgesasBought, int totalPricePerrosBought) {
-        int totalPrice = totalPriceHamburgesasBought + totalPricePerrosBought;
-        System.out.println("Total: " + totalPrice);
-    }
-
-    public static int takeMoreOrders(Scanner takeOrders){
-        System.out.println("Desea finalizar o agregar otro producto ?");
-        System.out.println("1. Agregar \n2. Finalizar");
-        return Integer.parseInt(takeOrders.nextLine());
-    }
-    private static int productToSell(Scanner takeOrders){
-        System.out.println("Seleccione producto: ");
-        System.out.println("1. Perro \n2. Hamburguesa");
-        return Integer.parseInt(takeOrders.nextLine());
-    }
-    private static int cuantityToSell(Scanner takeOrders){
-        System.out.println("Indique cantidad: ");
-        return Integer.parseInt(takeOrders.nextLine());
-    }
-
-    private static int sellProduct(int price, int cuantity){
-        return price*cuantity;
+        totalPriceHamburgesasBought = Product.sellProduct(hamburguesaPrice,hamburguesaCuantityToBuy);
+        totalPricePerrosBought = Product.sellProduct(perroPrice,perroCuantityToBuy);
+        Product.getTotalPrice(totalPriceHamburgesasBought,totalPricePerrosBought);
     }
 
 }
